# Conflicts:
#	CMakeLists.txt
#	appveyor.yml
#	src/command.c
#	src/command.h
#	src/d_clisrv.c
#	src/d_netcmd.c
#	src/dehacked.c
#	src/doomdef.h
#	src/hu_stuff.c
#	src/m_misc.c

+++ src/f_finale.c
@@ -175,7 +175,7 @@ static void F_SkyScroll(INT32 scrollspeed)
 	{
 		V_DrawFill(0, 0, BASEVIDWIDTH, BASEVIDHEIGHT, 120);
 
-		x = -animtimer;
+		x = -((INT32)animtimer);
 		y = 0;
 		while (x < w)
 		{
